add SET_LED_ON command
This allows gg to turn LED off without needing to know the color. This is handy when we are closing the app and we cant reliably request the UI to turn the LED off which is the source of truth for the color. This gives us a way to turn the LED off from the service rather than the UI