Change how the app starts
The UI needs to load for a short time in order for requestAnimationFrame to work and therefore ADS feature to work. by default, when the app loads on boot, it loads as hidden and the ADS feature will not work. This forces the UI to show for 500ms in hidden mode and that should be enough time to get the UI ready to support the ADS feature.